Move validate-schema to common
