+++ common/validate-schema.js
