+++ common/validate-schema.test.js
@@ -1,12 +1,22 @@
 /* eslint-env jest */
 'use strict';
-const { emailOnly } = require('./account-schemas');
+const Joi = require('@hapi/joi16');
 const validateSchema = require('./validate-schema');
 
 describe('validate schemas', () => {
     test('should have right error value if validations pass', () => {
+        const schema = Joi.object({
+            username: Joi.string()
+                .email()
+                .required()
+        });
+
+        const messages = {
+            username: [{ type: 'base', message: 'Invalid email' }]
+        };
+
         const validationResult = validateSchema(
-            emailOnly(),
+            { schema, messages },
             { username: 'test@test.com' }
         );
